Add mixin for to_dict() func
