+++ backend/python/app/models/__init__.py
@@ -12,7 +12,6 @@ def init_app(app):
     from .caregiver import Caregiver
     from .child_concern import ChildConcern
     from .daytime_contact import DaytimeContact
-    from .entity import Entity
     from .familial_concern import FamilialConcern
     from .transportation_method import TransportationMethod
     from .user import User
