+++ backend/python/app/models/base_mixin.py
@@ -1,0 +1,19 @@
+from sqlalchemy import inspect
+from sqlalchemy.orm.properties import ColumnProperty
+
+
+class BaseMixin(object):
+    def to_dict(self, include_relationships=False):
+        # define the address table
+        cls = type(self)
+
+        mapper = inspect(cls)
+        formatted = {}
+        for column in mapper.attrs:
+            field = column.key
+            attr = getattr(self, field)
+            if isinstance(column, ColumnProperty):
+                formatted[field] = attr
+            elif include_relationships:
+                formatted[field] = [obj.to_dict() for obj in attr]
+        return formatted
