+++ backend/python/app/models/child_concern.py
@@ -1,24 +1,8 @@
-from sqlalchemy import inspect
-from sqlalchemy.orm.properties import ColumnProperty
-
 from . import db
+from .base_mixin import BaseMixin
 
 
-class ChildConcern(db.Model):
+class ChildConcern(db.Model, BaseMixin):
     __tablename__ = "child_concerns"
     id = db.Column(db.Integer, primary_key=True, nullable=False)
     concern = db.Column(db.String, nullable=False)
-
-    def to_dict(self, include_relationships=False):
-        cls = type(self)
-
-        mapper = inspect(cls)
-        formatted = {}
-        for column in mapper.attrs:
-            field = column.key
-            attr = getattr(self, field)
-            if isinstance(column, ColumnProperty):
-                formatted[field] = attr
-            elif include_relationships:
-                formatted[field] = [obj.to_dict() for obj in attr]
-        return formatted
