--- backend/python/app/models/enum.py
@@ -1,3 +1,0 @@
-from . import db
-
-enum = db.Enum("A", "B", "C", "D", name="enum")
