+++ backend/python/app/models/transportation_method.py
@@ -1,24 +1,8 @@
-from sqlalchemy import inspect
-from sqlalchemy.orm.properties import ColumnProperty
-
 from . import db
+from .base_mixin import BaseMixin
 
 
-class TransportationMethod(db.Model):
+class TransportationMethod(db.Model, BaseMixin):
     __tablename__ = "transportation_methods"
     id = db.Column(db.Integer, primary_key=True, nullable=False)
     transportation_method = db.Column(db.String, nullable=False)
-
-    def to_dict(self, include_relationships=False):
-        cls = type(self)
-
-        mapper = inspect(cls)
-        formatted = {}
-        for column in mapper.attrs:
-            field = column.key
-            attr = getattr(self, field)
-            if isinstance(column, ColumnProperty):
-                formatted[field] = attr
-            elif include_relationships:
-                formatted[field] = [obj.to_dict() for obj in attr]
-        return formatted
